enertrack/requirements.txt
Flask
flask-mysqldb
gunicorn
cryptography
paho-mqtt
influxdb-client
aiomqtt
certifi
python-telegram-bot
requests
APScheduler
pytz
mysql-connector-python
